Make movement while confused tend more towards the chosen direction
Gives movement while confused a smaller chance of being fully random, and
otherwise moves either in the chosen direction or one of the adjacent
directions. The end result is a very slightly lower chance to move in the
precise chosen direction, but a pretty significant chance to move roughly
towards the target.